Added basic HTML output format.
One of the features I was missing when Yesterday I tried this tool out.
Now it is implemented with minor features and basic style, but may be
expended easily by updating the generated CSS file.

First I thougth source attachment as hyperlinks would be cool, but I
realized this is not always a feasible goal (see my TODO comment in the
code).

Hope you will be able to use this code for something.