Added styled HTML output generation.
The HTML generation now may include templates files direclty. Two templates
were supported as an example: one for the CSS and a simple HTML for the
general information in the header about the various metrics.

It would be really nice to update the sample CSS I made to the one used at
the official site at: http://www.clarkware.com/software/JDepend.html